requirements.txt
attrs==22.1.0
cycler==0.11.0
fonttools==4.37.1
iniconfig==1.1.1
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
matplotlib==3.5.3
mccabe==0.6.1
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pymongo==4.2.0
pyparsing==2.4.5
pytest==7.1.3
python-dateutil==2.8.1
six==1.13.0
stop-words==2018.7.23
tomli==2.0.1
wordcloud==1.8.2.2
wrapt==1.11.2
